Audio changes
Added a few sound effects to the game when the brick or paddle is hit by a ball and when the paddle is hit by a powerUp. Separated Music and SFX in audio settings in menu. Fixed minor things: ball random trajectory force on x on the start is smaller; normal paddle speed is 35.